% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_code}
\alias{format_code}
\alias{format_file}
\title{Format R code}
\usage{
format_code(code)

format_file(file)
}
\arguments{
\item{code}{character scalar containing R code}

\item{file}{path of R file to format}
}
\value{
\code{format_code} returns the formatted R code. \code{format_file} returns
a boolean that indicates if the file was changed
}
\description{
Formats a string containing R code.
}
\examples{
bad_code <- "x+ 1  *(3 -0 )"
format_code(bad_code)

tmp <- tempfile(fileext = ".R")
writeLines(tmp, bad_code)
format_file(tmp)
readLines(tmp

}
