% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rair-package.R
\docType{package}
\name{rair-package}
\alias{rair}
\alias{rair-package}
\title{rair: What the Package Does (One Line, Title Case)}
\description{
Format R code with Air, an R formatter written in Rust.
}
\author{
\strong{Maintainer}: Mark Paulin \email{paulin.mark@googlemail.com}

}
\keyword{internal}
